Moved export_icon into separate file
