+++ export_icon.py
@@ -1,0 +1,41 @@
+def export_icon(icon, size, filename, font, color):
+    image = Image.new("RGBA", (size, size), color=(0,0,0,0))
+
+    draw = ImageDraw.Draw(image)
+
+    # Initialize font
+    font = ImageFont.truetype(font, size)
+
+    # Determine the dimensions of the icon
+    width,height = draw.textsize(icons[icon], font=font)
+
+    draw.text(((size - width) / 2, (size - height) / 2), icons[icon],
+            font=font, fill=color)
+
+    # Get bounding box
+    bbox = image.getbbox()
+
+    # Create an alpha mask
+    imagemask = Image.new("L", (size, size), 0)
+    drawmask = ImageDraw.Draw(imagemask)
+
+    # Draw the icon on the mask
+    drawmask.text(((size - width) / 2, (size - height) / 2), icons[icon],
+        font=font, fill=255)
+
+    # Create a solid color image and apply the mask
+    iconimage = Image.new("RGBA", (size,size), color)
+    iconimage.putalpha(imagemask)
+
+    if bbox:
+        iconimage = iconimage.crop(bbox)
+
+    borderw = int((size - (bbox[2] - bbox[0])) / 2)
+    borderh = int((size - (bbox[3] - bbox[1])) / 2)
+
+    # Create output image
+    outimage = Image.new("RGBA", (size, size), (0,0,0,0))
+    outimage.paste(iconimage, (borderw,borderh))
+
+    # Save file
+    outimage.save(filename)
